% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia.pal.R
\name{arcadia.pal}
\alias{arcadia.pal}
\title{Generate colors using Aracadia Science's color palettes}
\usage{
arcadia.pal(
  n = 3,
  name = c("Neutral", "Accent", "Lighter_accents"),
  choose_random = FALSE
)
}
\arguments{
\item{n}{Number of colors.}

\item{name}{Which palette to use.}

\item{choose_random}{Randomly n sample colors from palette.}
}
\description{
Generate colors using Aracadia Science's color palettes
}
\examples{
arcadia.pal(n = 3, name = 'Neutral')
arcadia.pal(n = 5, name = 'Neutral', choose_random = TRUE)
}
