% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darken_color.R
\name{darken_color}
\alias{darken_color}
\title{Darken colors by a desired amount}
\usage{
darken_color(color, factor = 1.4)
}
\arguments{
\item{color}{Color(s) to darken.}

\item{factor}{Factor to darken by.}
}
\description{
Darken colors by a desired amount
}
\examples{
darken_color('blue', factor = 1.5)
darken_color(arcadia.pal(n = 3, name = 'Neutral'))
}
